// sdk/durabletask/arm-durabletask/src/models/models.ts
export interface Resource {
  /** Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName} */
  readonly id?: string;
  /** The name of the resource */
  readonly name?: string;
  /** The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts" */
  readonly type?: string;
  /** Azure Resource Manager metadata containing createdBy and modifiedBy information. */
  readonly systemData?: SystemData;
}
// ...
export interface TrackedResource extends Resource {
  /** Resource tags. */
  tags?: Record<string, string>;
  /** The geo-location where the resource lives */
  location: string;
}

export function trackedResourceSerializer(
  item: TrackedResource,
): Record<string, unknown> {
  return {
    tags: !item.tags ? item.tags : (serializeRecord(item.tags as any) as any),
    location: item["location"],
  };
}

/** A Durable Task Scheduler resource */
export interface Scheduler extends TrackedResource {
  /** The resource-specific properties for this resource. */
  properties?: SchedulerProperties;
}

export function schedulerSerializer(item: Scheduler): Record<string, unknown> {
  return {
    tags: !item.tags ? item.tags : (serializeRecord(item.tags as any) as any),
    location: item["location"],
    properties: !item.properties
      ? item.properties
      : schedulerPropertiesSerializer(item.properties),
  };
}

/** Details of the Scheduler */
export interface SchedulerProperties {
  /** The status of the last operation */
  readonly provisioningState?: ProvisioningState;
  /** URL of the durable task scheduler */
  readonly endpoint?: string;
  /** IP allow list for durable task scheduler. Values can be IPv4, IPv6 or CIDR */
  ipAllowlist: string[];
  /** SKU of the durable task scheduler */
  sku: SchedulerSku;
}

export function schedulerPropertiesSerializer(
  item: SchedulerProperties,
): Record<string, unknown> {
  return {
    ipAllowlist: item["ipAllowlist"],
    sku: schedulerSkuSerializer(item.sku),
  };
}

/** The SKU (Stock Keeping Unit) assigned to this durable task scheduler */
export interface SchedulerSku {
  /** The name of the SKU */
  name: string;
  /** The SKU capacity. This allows scale out/in for the resource and impacts zone redundancy */
  capacity?: number;
  /** Indicates whether the current SKU configuration is zone redundant */
  readonly redundancyState?: RedundancyState;
}
// ...
export interface _SchedulerListResult {
  /** The Scheduler items on this page */
  value: Scheduler[];
  /** The link to the next page of items */
  nextLink?: string;
}
// ...
export type ProvisioningState =
  | string
  | ResourceProvisioningState
  | "Provisioning"
  | "Updating"
  | "Deleting"
  | "Accepted";